requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
prometheus-client==0.19.0
psutil==5.9.6
python-dotenv==1.0.0
pytest==7.4.3
httpx>=0.24.0,<0.25.0
redis==5.0.1
boto3==1.29.3
tenacity==8.2.3
psycopg2-binary==2.9.9
rich==13.7.0
supabase==2.3.0
aiohttp==3.9.3
aioredis==2.0.1
pydantic==2.10.6
pyyaml==6.0.1
torch==2.1.0
transformers==4.35.2
diffusers==0.23.0
accelerate==0.27.2
huggingface-hub==0.20.3
